Refactored qreal, QByteArray, QString, and QUrl into different files
