+++ qttypes/src/core/mod.rs
@@ -1,0 +1,9 @@
+mod primitives;
+mod qbytearray;
+mod qstring;
+mod qurl;
+
+pub use self::primitives::qreal;
+pub use self::qbytearray::QByteArray;
+pub use self::qstring::QString;
+pub use self::qurl::QUrl;
