+++ qttypes/src/core/qbytearray.rs
@@ -1,0 +1,83 @@
+#[cfg(not(no_qt))]
+use cpp::{cpp, cpp_class};
+
+use std::fmt::Display;
+use std::os::raw::c_char;
+use std::str::Utf8Error;
+
+use crate::core::QString;
+
+cpp! {{
+    #include <QtCore/QString>
+    #include <QtCore/QByteArray>
+}}
+
+cpp_class!(
+    /// Wrapper around [`QByteArray`][class] class.
+    ///
+    /// [class]: https://doc.qt.io/qt-5/qbytearray.html
+    #[derive(PartialEq, PartialOrd, Eq, Ord)]
+    pub unsafe struct QByteArray as "QByteArray"
+);
+impl QByteArray {
+    pub fn to_slice(&self) -> &[u8] {
+        unsafe {
+            let mut size: usize = 0;
+            let c_ptr = cpp!([self as "const QByteArray*", mut size as "size_t"] -> *const u8 as "const char*" {
+                size = self->size();
+                return self->constData();
+            });
+            std::slice::from_raw_parts(c_ptr, size)
+        }
+    }
+    pub fn to_str(&self) -> Result<&str, Utf8Error> {
+        std::str::from_utf8(self.to_slice())
+    }
+}
+impl<'a> From<&'a [u8]> for QByteArray {
+    /// Constructs a `QByteArray` from a slice. (Copy the slice.)
+    fn from(s: &'a [u8]) -> QByteArray {
+        let len = s.len();
+        let ptr = s.as_ptr();
+        cpp!(unsafe [len as "size_t", ptr as "char*"] -> QByteArray as "QByteArray" {
+            return QByteArray(ptr, len);
+        })
+    }
+}
+impl<'a> From<&'a str> for QByteArray {
+    /// Constructs a `QByteArray` from a `&str`. (Copy the string.)
+    fn from(s: &'a str) -> QByteArray {
+        s.as_bytes().into()
+    }
+}
+impl From<String> for QByteArray {
+    /// Constructs a `QByteArray` from a `String`. (Copy the string.)
+    fn from(s: String) -> QByteArray {
+        QByteArray::from(&*s)
+    }
+}
+impl From<QString> for QByteArray {
+    /// Converts a `QString` to a `QByteArray`
+    fn from(s: QString) -> QByteArray {
+        cpp!(unsafe [s as "QString"] -> QByteArray as "QByteArray" {
+            return std::move(s).toUtf8();
+        })
+    }
+}
+impl Display for QByteArray {
+    /// Prints the contents of the `QByteArray` if it contains UTF-8, do nothing otherwise.
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        unsafe {
+            let c_ptr = cpp!([self as "const QByteArray*"] -> *const c_char as "const char*" {
+                return self->constData();
+            });
+            f.write_str(std::ffi::CStr::from_ptr(c_ptr).to_str().map_err(|_| Default::default())?)
+        }
+    }
+}
+impl std::fmt::Debug for QByteArray {
+    /// Prints the contents of the `QByteArray` if it contains UTF-8,  nothing otherwise
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self)
+    }
+}
