+++ qttypes/src/lib.rs
@@ -120,19 +120,22 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 #![cfg_attr(no_qt, allow(unused))]
 
+use std::collections::HashMap;
 use std::convert::From;
 use std::fmt::Display;
 use std::iter::FromIterator;
 use std::ops::{Index, IndexMut};
 use std::os::raw::c_char;
 use std::str::Utf8Error;
-use std::collections::HashMap;
 
 #[cfg(feature = "chrono")]
 use chrono::prelude::*;
 
 #[cfg(not(no_qt))]
 use cpp::{cpp, cpp_class};
+
+mod core;
+pub use crate::core::{qreal, QByteArray, QString, QUrl};
 
 #[cfg(no_qt)]
 mod no_qt {
@@ -175,76 +178,6 @@ cpp! {{
     #include <QtGui/QPen>
     #include <QtGui/QBrush>
 }}
-
-cpp_class!(
-    /// Wrapper around [`QByteArray`][class] class.
-    ///
-    /// [class]: https://doc.qt.io/qt-5/qbytearray.html
-    #[derive(PartialEq, PartialOrd, Eq, Ord)]
-    pub unsafe struct QByteArray as "QByteArray"
-);
-impl QByteArray {
-    pub fn to_slice(&self) -> &[u8] {
-        unsafe {
-            let mut size: usize = 0;
-            let c_ptr = cpp!([self as "const QByteArray*", mut size as "size_t"] -> *const u8 as "const char*" {
-                size = self->size();
-                return self->constData();
-            });
-            std::slice::from_raw_parts(c_ptr, size)
-        }
-    }
-    pub fn to_str(&self) -> Result<&str, Utf8Error> {
-        std::str::from_utf8(self.to_slice())
-    }
-}
-impl<'a> From<&'a [u8]> for QByteArray {
-    /// Constructs a `QByteArray` from a slice. (Copy the slice.)
-    fn from(s: &'a [u8]) -> QByteArray {
-        let len = s.len();
-        let ptr = s.as_ptr();
-        cpp!(unsafe [len as "size_t", ptr as "char*"] -> QByteArray as "QByteArray" {
-            return QByteArray(ptr, len);
-        })
-    }
-}
-impl<'a> From<&'a str> for QByteArray {
-    /// Constructs a `QByteArray` from a `&str`. (Copy the string.)
-    fn from(s: &'a str) -> QByteArray {
-        s.as_bytes().into()
-    }
-}
-impl From<String> for QByteArray {
-    /// Constructs a `QByteArray` from a `String`. (Copy the string.)
-    fn from(s: String) -> QByteArray {
-        QByteArray::from(&*s)
-    }
-}
-impl From<QString> for QByteArray {
-    /// Converts a `QString` to a `QByteArray`
-    fn from(s: QString) -> QByteArray {
-        cpp!(unsafe [s as "QString"] -> QByteArray as "QByteArray" {
-            return std::move(s).toUtf8();
-        })
-    }
-}
-impl Display for QByteArray {
-    /// Prints the contents of the `QByteArray` if it contains UTF-8, do nothing otherwise.
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        unsafe {
-            let c_ptr = cpp!([self as "const QByteArray*"] -> *const c_char as "const char*" {
-                return self->constData();
-            });
-            f.write_str(std::ffi::CStr::from_ptr(c_ptr).to_str().map_err(|_| Default::default())?)
-        }
-    }
-}
-impl std::fmt::Debug for QByteArray {
-    /// Prints the contents of the `QByteArray` if it contains UTF-8,  nothing otherwise
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", self)
-    }
-}
 
 cpp_class!(
     /// Wrapper around [`QDate`][class] class.
@@ -474,7 +407,7 @@ impl QDateTime {
         #else
                     return QDateTime(date);
         #endif
-                })
+        })
     }
 
     /// Wrapper around [`QDateTime(const QDate &date, const QTime &time, Qt::TimeSpec spec = Qt::LocalTime)`][ctor] constructor.
@@ -576,96 +509,6 @@ fn test_qdatetime_is_valid() {
     let invalid_qdatetime_from_invalid_date_invalid_time =
         QDateTime::from_date_time_local_timezone(invalid_qdate, invalid_qtime);
     assert!(!invalid_qdatetime_from_invalid_date_invalid_time.is_valid());
-}
-
-cpp_class!(
-    /// Wrapper around [`QUrl`][class] class.
-    ///
-    /// [class]: https://doc.qt.io/qt-5/qurl.html
-    #[derive(PartialEq, PartialOrd, Eq, Ord)]
-    pub unsafe struct QUrl as "QUrl"
-);
-impl QUrl {
-    /// Wrapper around [`QUrl::fromUserInput(const QString &userInput)`][method] static method.
-    ///
-    /// [method]: https://doc.qt.io/qt-5/qurl.html#fromUserInput
-    pub fn from_user_input(user_input: QString) -> QUrl {
-        cpp!(unsafe [user_input as "QString"] -> QUrl as "QUrl" {
-            return QUrl::fromUserInput(user_input);
-        })
-    }
-}
-impl From<QString> for QUrl {
-    fn from(qstring: QString) -> QUrl {
-        cpp!(unsafe [qstring as "QString"] -> QUrl as "QUrl" {
-            return QUrl(qstring);
-        })
-    }
-}
-
-cpp_class!(
-    /// Wrapper around [`QString`][class] class.
-    ///
-    /// [class]: https://doc.qt.io/qt-5/qstring.html
-    #[derive(PartialEq, PartialOrd, Eq, Ord)]
-    pub unsafe struct QString as "QString"
-);
-impl QString {
-    /// Return a slice containing the UTF-16 data.
-    pub fn to_slice(&self) -> &[u16] {
-        unsafe {
-            let mut size: usize = 0;
-            let c_ptr = cpp!([self as "const QString*", mut size as "size_t"] -> *const u16 as "const QChar*" {
-                size = self->size();
-                return self->constData();
-            });
-            std::slice::from_raw_parts(c_ptr, size)
-        }
-    }
-}
-impl From<QUrl> for QString {
-    /// Wrapper around [`QUrl::toString(QUrl::FormattingOptions=...)`][method] method.
-    ///
-    /// # Wrapper-specific
-    ///
-    /// Formatting options are left at defaults.
-    ///
-    /// [method]: https://doc.qt.io/qt-5/qurl.html#toString
-    fn from(qurl: QUrl) -> QString {
-        cpp!(unsafe [qurl as "QUrl"] -> QString as "QString" {
-            return qurl.toString();
-        })
-    }
-}
-impl<'a> From<&'a str> for QString {
-    /// Copy the data from a `&str`.
-    fn from(s: &'a str) -> QString {
-        let len = s.len();
-        let ptr = s.as_ptr();
-        cpp!(unsafe [len as "size_t", ptr as "char*"] -> QString as "QString" {
-            return QString::fromUtf8(ptr, len);
-        })
-    }
-}
-impl From<String> for QString {
-    fn from(s: String) -> QString {
-        QString::from(&*s)
-    }
-}
-impl Into<String> for QString {
-    fn into(self) -> String {
-        String::from_utf16_lossy(self.to_slice())
-    }
-}
-impl Display for QString {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        QByteArray::from(self.clone()).fmt(f)
-    }
-}
-impl std::fmt::Debug for QString {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", self)
-    }
 }
 
 cpp_class!(
@@ -1062,17 +905,6 @@ impl QModelIndex {
     }
 }
 
-/// Bindings for [`qreal`][type] typedef.
-///
-/// [type]: https://doc.qt.io/qt-5/qtglobal.html#qreal-typedef
-#[allow(non_camel_case_types)]
-#[cfg(qreal_is_float)]
-pub type qreal = f32;
-
-#[allow(non_camel_case_types)]
-#[cfg(not(qreal_is_float))]
-pub type qreal = f64;
-
 /// Bindings for [`QRectF`][class] class.
 ///
 /// [class]: https://doc.qt.io/qt-5/qrectf.html
@@ -1416,7 +1248,7 @@ pub enum PenStyle {
     DotLine = 3,
     DashDotLine = 4,
     DashDotDotLine = 5,
-    CustomDashLine = 6
+    CustomDashLine = 6,
 }
 cpp_class!(
     /// Wrapper around [`QPen`][class] class.
@@ -1446,29 +1278,27 @@ impl QPen {
         cpp!(unsafe [self as "QPen*", width as "qreal"] { return self->setWidthF(width); });
     }
 
-//    QBrush	brush() const
-//    Qt::PenCapStyle	capStyle() const
-//    QColor	color() const
-//    qreal	dashOffset() const
-//    QVector<qreal>	dashPattern() const
-//    bool	isCosmetic() const
-//    bool	isSolid() const
-//    Qt::PenJoinStyle	joinStyle() const
-//    qreal	miterLimit() const
-//    void	setBrush(const QBrush &brush)
-//    void	setCapStyle(Qt::PenCapStyle style)
-//    void	setCosmetic(bool cosmetic)
-//    void	setDashOffset(qreal offset)
-//    void	setDashPattern(const QVector<qreal> &pattern)
-//    void	setJoinStyle(Qt::PenJoinStyle style)
-//    void	setMiterLimit(qreal limit)
-//    Qt::PenStyle	style() const
-//    void	swap(QPen &other)
-//    int	width() const
-//    qreal	widthF() const
-
+    //    QBrush	brush() const
+    //    Qt::PenCapStyle	capStyle() const
+    //    QColor	color() const
+    //    qreal	dashOffset() const
+    //    QVector<qreal>	dashPattern() const
+    //    bool	isCosmetic() const
+    //    bool	isSolid() const
+    //    Qt::PenJoinStyle	joinStyle() const
+    //    qreal	miterLimit() const
+    //    void	setBrush(const QBrush &brush)
+    //    void	setCapStyle(Qt::PenCapStyle style)
+    //    void	setCosmetic(bool cosmetic)
+    //    void	setDashOffset(qreal offset)
+    //    void	setDashPattern(const QVector<qreal> &pattern)
+    //    void	setJoinStyle(Qt::PenJoinStyle style)
+    //    void	setMiterLimit(qreal limit)
+    //    Qt::PenStyle	style() const
+    //    void	swap(QPen &other)
+    //    int	width() const
+    //    qreal	widthF() const
 }
-
 
 /// Bindings for [`Qt::BrushStyle`][enum] enum.
 ///
@@ -1518,7 +1348,6 @@ impl QBrush {
         cpp!(unsafe [self as "QBrush*", style as "Qt::BrushStyle"] { return self->setStyle(style); });
     }
 }
-
 
 /// Bindings for [`QLineF`][class] class.
 ///
@@ -1577,17 +1406,27 @@ impl QPainter {
             self->drawImage(point, image);
         });
     }
-    pub fn draw_image_fit_rect_with_source(&mut self, rectangle: QRectF, image: QImage, source_rect: QRectF) {
+    pub fn draw_image_fit_rect_with_source(
+        &mut self,
+        rectangle: QRectF,
+        image: QImage,
+        source_rect: QRectF,
+    ) {
         cpp!(unsafe [self as "QPainter*", rectangle as "QRectF", image as "QImage", source_rect as "QRectF"] {
             self->drawImage(rectangle, image, source_rect);
         });
     }
-    pub fn draw_image_at_point_with_source(&mut self, point: QPointF, image: QImage, source_rect: QRectF) {
+    pub fn draw_image_at_point_with_source(
+        &mut self,
+        point: QPointF,
+        image: QImage,
+        source_rect: QRectF,
+    ) {
         cpp!(unsafe [self as "QPainter*", point as "QPointF", image as "QImage", source_rect as "QRectF"] {
             self->drawImage(point, image, source_rect);
         });
     }
-    
+
     pub fn draw_line(&mut self, line: QLineF) {
         cpp!(unsafe [self as "QPainter*", line as "QLineF"] {
             self->drawLine(line);
@@ -1678,7 +1517,7 @@ impl QPainter {
             self->eraseRect(rectangle);
         });
     }
-    
+
     pub fn fill_rect(&mut self, rectangle: QRectF, brush: QBrush) {
         cpp!(unsafe [self as "QPainter*", rectangle as "QRectF", brush as "QBrush"] {
             self->fillRect(rectangle, brush);
@@ -1750,9 +1589,9 @@ impl QPainter {
         });
     }
 
-// void	setBackgroundMode(Qt::BGMode mode)
-// void	setCompositionMode(QPainter::CompositionMode mode)
-// void	setFont(const QFont &font)
+    // void	setBackgroundMode(Qt::BGMode mode)
+    // void	setCompositionMode(QPainter::CompositionMode mode)
+    // void	setFont(const QFont &font)
 }
 
 /// Bindings for [`QPainter::RenderHint`][enum] enum.
@@ -1868,13 +1707,13 @@ fn test_qjsonvalue() {
     let test_f64_variant2 = QVariant::from(1.2345);
     assert!(test_f64_variant == test_f64_variant2);
     assert_eq!(<QJsonValue as Into<f64>>::into(test_f64), 1.2345);
-    
+
     let values = QJsonArray::from(vec![
-        QJsonValue::from(QString::from("test")), 
-        QJsonValue::from(true), 
-        QJsonValue::from(false), 
-        QJsonValue::from(1.2345), 
-        QJsonValue::from(456.0), 
+        QJsonValue::from(QString::from("test")),
+        QJsonValue::from(true),
+        QJsonValue::from(false),
+        QJsonValue::from(1.2345),
+        QJsonValue::from(456.0),
     ]);
 
     assert_eq!(values.to_json().to_string(), "[\"test\",true,false,1.2345,456]");
@@ -2006,13 +1845,12 @@ fn test_qjsonobject() {
     let object = QJsonObject::from(hashmap);
     assert_eq!(object.to_json().to_string(), "{\"key\":\"value\",\"test\":\"hello\"}");
 
-
     let array = QJsonArray::from(vec![
-        QJsonValue::from(QString::from("test")), 
-        QJsonValue::from(true), 
-        QJsonValue::from(false), 
-        QJsonValue::from(1.2345), 
-        QJsonValue::from(456.0), 
+        QJsonValue::from(QString::from("test")),
+        QJsonValue::from(true),
+        QJsonValue::from(false),
+        QJsonValue::from(1.2345),
+        QJsonValue::from(456.0),
     ]);
 
     let mut valuemap = HashMap::new();
